revert to old behavior of error reporting in ScopeRunsInVisitor
Apparently the checker framework can't report errors in other compilation
units. It will instead report an error in the currently checked file at the
same line number. This fix goes back to the old behavior, where classes with
source are double checked, but errors on the class are only reported when they
are properly visited, not when visited by necessity/use in another class.